# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Oxidrop Firewall API",
    "description": "Core API for the Oxidrop firewall system",
    "license": {
      "name": "GPL-2.0-only",
      "identifier": "GPL-2.0-only"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/users/create_user": {
      "post": {
        "tags": [
          "crate::api"
        ],
        "operationId": "create_user_endpoint",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "User already exists"
          }
        },
        "security": [
          {
            "cookie_auth": []
          }
        ]
      }
    },
    "/api/v1/users/delete_user": {
      "delete": {
        "tags": [
          "crate::api"
        ],
        "operationId": "delete_user_endpoint",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteUserReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "User deleted"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "cookie_auth": []
          }
        ]
      }
    },
    "/api/v1/users/modify_user": {
      "put": {
        "tags": [
          "crate::api"
        ],
        "operationId": "modify_user_endpoint",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModifyUserReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User modified"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "cookie_auth": []
          }
        ]
      }
    },
    "/api/v1/users/rename_user": {
      "patch": {
        "tags": [
          "crate::api"
        ],
        "operationId": "rename_user_endpoint",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameUserReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User renamed"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "cookie_auth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateUserReq": {
        "type": "object",
        "required": [
          "username",
          "password",
          "role",
          "permissions"
        ],
        "properties": {
          "password": {
            "type": "string"
          },
          "permissions": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "role": {
            "$ref": "#/components/schemas/RolesUser"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "DeleteUserReq": {
        "type": "object",
        "required": [
          "target_username"
        ],
        "properties": {
          "target_username": {
            "type": "string"
          }
        }
      },
...
      "ModifyUserReq": {
        "type": "object",
        "required": [
          "target_username",
          "role",
          "permissions",
          "is_active"
        ],
        "properties": {
          "is_active": {
            "type": "integer",
            "format": "int64"
          },
          "permissions": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "role": {
            "$ref": "#/components/schemas/RolesUser"
          },
          "target_username": {
            "type": "string"
          }
        }
      },
...
      "RenameUserReq": {
        "type": "object",
        "required": [
          "current_username",
          "new_username"
        ],
        "properties": {
          "current_username": {
            "type": "string"
          },
          "new_username": {
            "type": "string"
          }
        }
      },
      "RolesUser": {
        "type": "string",
        "enum": [
          "Viewer",
          "Admin"
        ]
      },
...
    }
  }
}